+++ src/lib/services/validator.ts
@@ -1,9 +1,11 @@
 import type { Coin } from "@cosmjs/stargate";
 import axios from "axios";
+import { z } from "zod";
 
 import { CURR_THEME } from "env";
-import type { StakingShare, Validator, ValidatorAddr } from "lib/types";
-import { removeSpecialChars } from "lib/utils";
+import { BlockVote, zBig, zUtcDate, zValidatorData } from "lib/types";
+import type { Option, StakingShare, Validator, ValidatorAddr } from "lib/types";
+import { parseWithError, removeSpecialChars, snakeToCamel } from "lib/utils";
 
 interface ValidatorResponse {
   operator_address: ValidatorAddr;
@@ -74,3 +76,89 @@ export const resolveValIdentity = async (
       })
   );
 };
+
+const zValidatorsResponse = z
+  .object({
+    items: z.array(zValidatorData),
+    total: z.number().nonnegative(),
+    total_voting_power: zBig,
+  })
+  .transform(snakeToCamel);
+export type ValidatorsResponse = z.infer<typeof zValidatorsResponse>;
+
+export const getValidators = async (
+  endpoint: string,
+  limit: number,
+  offset: number,
+  isActive: boolean,
+  sortBy: string,
+  isDesc: boolean,
+  search: Option<string>
+) =>
+  axios
+    .get(`${endpoint}`, {
+      params: {
+        limit,
+        offset,
+        is_active: isActive,
+        sort_by: sortBy,
+        is_desc: isDesc,
+        search,
+      },
+    })
+    .then(({ data }) => parseWithError(zValidatorsResponse, data));
+
+const zValidatorDataResponse = z
+  .object({
+    info: zValidatorData.nullable(),
+    self_voting_power: zBig,
+    total_voting_power: zBig,
+  })
+  .transform(snakeToCamel);
+export type ValidatorDataResponse = z.infer<typeof zValidatorDataResponse>;
+
+export const getValidatorData = async (
+  endpoint: string,
+  validatorAddress: ValidatorAddr
+) =>
+  axios
+    .get(`${endpoint}/${encodeURIComponent(validatorAddress)}/info`)
+    .then(({ data }) => parseWithError(zValidatorDataResponse, data));
+
+const zValidatorUptimeResponse = z
+  .object({
+    uptime: z.object({
+      signed_block: z.number(),
+      proposed_block: z.number(),
+      missed_blocks: z.number(),
+      total: z.number(),
+    }),
+    recent_blocks: z
+      .object({
+        height: z.number(),
+        vote: z.nativeEnum(BlockVote),
+      })
+      .array(),
+    events: z
+      .object({
+        height: z.number(),
+        timestamp: zUtcDate,
+        is_jailed: z.boolean(),
+      })
+      .array(),
+  })
+  .transform(snakeToCamel);
+export type ValidatorUptimeResponse = z.infer<typeof zValidatorUptimeResponse>;
+
+export const getValidatorUptime = async (
+  endpoint: string,
+  validatorAddress: ValidatorAddr,
+  blocks: number
+) =>
+  axios
+    .get(`${endpoint}/${encodeURIComponent(validatorAddress)}/uptime`, {
+      params: {
+        blocks,
+      },
+    })
+    .then(({ data }) => parseWithError(zValidatorUptimeResponse, data));